- Moved all helper functions in utils.lua - Created init.lua to load all FRE extensions - Adjusted drag code to use utils
