+++ lua/ge/extensions/gameplay/drag/dragTypes/dragPracticeRace.lua
@@ -9,6 +9,7 @@ local dGeneral, dUtils
 local dragData
 local logTag = ""
 local freeroamEvents = require("gameplay/events/freeroamEvents")
+local freeroamUtils = require("gameplay/events/freeroam/utils")
 local hasActivityStarted = false
 local function onExtensionLoaded()
   --log("I", logTag, "dragRace extension loaded")
@@ -79,13 +80,13 @@ local function onUpdate(dtReal, dtSim, dtRaw)
       if phase.completed and not racer.isFinished then
         log('I', logTag, 'Racer: '.. racer.vehId ..' completed phase: '.. phase.name)
         if phase.name == "stage" then
-          freeroamEvents.displayStagedMessage("drag")
+          freeroamUtils.displayStagedMessage("drag")
         elseif phase.name == "countdown" then
-          freeroamEvents.displayMessage(freeroamEvents.getStartMessage("drag"), 5)
-          freeroamEvents.saveAndSetTrafficAmount(0)
+          freeroamUtils.displayStartMessage("drag")
+          freeroamUtils.saveAndSetTrafficAmount(0)
         elseif phase.name == "race" then
           freeroamEvents.payoutDragRace("drag", racer.timers.time_1_4.value, racer.vehSpeed * 2.2369362921)
-          freeroamEvents.restoreTrafficAmount()
+          freeroamUtils.restoreTrafficAmount()
         end
         dUtils.changeRacerPhase(racer)
       end
